feature: add main_train_network.py

+++ datasets/get_dataset_v1.py
@@ -13,8 +13,8 @@ def get_dataset_v1(N: int, K: float, u_bar: float, R: float) -> tuple[np.ndarray
         y = uniform(-limit, limit)
         theta = uniform(0, 2 * math.pi)
         inputs.append([x, y, math.cos(theta), math.sin(theta)])
-        outputs.append([1, -1])
-    return np.array(inputs), np.array(outputs)
+        outputs.append([1, -1.])
+    return np.array(inputs,dtype=float), np.array(outputs, dtype=float)
 
 if __name__ == '__main__':
     inputs, outputs = get_dataset_v1(10, 2, 2, 2)
